pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imapclient>=3.1.0",
    "ollama>=0.4",
    "pandas>=2.3.3",
    "playwright>=1.58.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "streamlit>=1.55.0",
]
